+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Snackbar for user actions [#51](https://github.com/IN-CORE/incore-studio/issues/51)
 - Create empty project [#42](https://github.com/IN-CORE/incore-studio/issues/42)
 - Modal that adds resources from service to project [#40](https://github.com/IN-CORE/incore-studio/issues/40)
+- Modal to create a new workflow and add to project [#41](https://github.com/IN-CORE/incore-studio/issues/41)
 
 ### Changed